pyproject.toml
[tool.poetry]
name = "crypto"
version = "0.1.0"
description = ""
authors = ["Volodya Orlov <volodka.orlov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dependency-injector = "^4.37.0"
requests = "^2.26.0"
pydantic = "^1.8.2"
progress = "^1.6"
tqdm = "^4.62.3"
PyNaCl = "^1.4.0"
fastapi = "^0.70.1"
SQLAlchemy = "^1.4.28"
PyYAML = "^6.0"
uvicorn = "^0.16.0"
aiosqlite = "^0.17.0"
Jinja2 = "^3.0.3"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
